requirements.txt
requests==2.31.0
snowflake-connector-python==3.7.0